daily-job-bot/requirements.txt
requests
openai
numpy
perplexityai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib